+++ _Typings/utils/guid.d.ts
@@ -2,7 +2,7 @@
 // For details: https://github.com/Withered-Flower-0422/BST/blob/main/LICENSE
 
 declare module "guid" {
-    type IsHexStringOfExactLength<S extends string, L extends number, C extends any[] = []> = C["length"] extends L
+    type IsHexStringOfExactLength<S extends string, L extends number, C extends 0[] = []> = C["length"] extends L
         ? S extends ""
             ? true
             : false
@@ -23,43 +23,36 @@ declare module "guid" {
               | "d"
               | "e"
               | "f"}${infer R}`
-        ? IsHexStringOfExactLength<R, L, [...C, any]>
+        ? IsHexStringOfExactLength<R, L, [...C, 0]>
         : false
 
-    type ValidateGuidParts<
-        P1 extends string,
-        P2 extends string,
-        P3 extends string,
-        P4 extends string,
-        P5 extends string
-    > = IsHexStringOfExactLength<P1, 8> extends true
-        ? IsHexStringOfExactLength<P2, 4> extends true
-            ? IsHexStringOfExactLength<P3, 4> extends true
-                ? IsHexStringOfExactLength<P4, 4> extends true
-                    ? IsHexStringOfExactLength<P5, 12> extends true
-                        ? true
-                        : false
-                    : false
-                : false
+    type ValidateGuidParts<P extends [string, string, string, string, string]> = P extends [
+        infer P1 extends string,
+        infer P2 extends string,
+        infer P3 extends string,
+        infer P4 extends string,
+        infer P5 extends string
+    ]
+        ? [
+              IsHexStringOfExactLength<P1, 8>,
+              IsHexStringOfExactLength<P2, 4>,
+              IsHexStringOfExactLength<P3, 4>,
+              IsHexStringOfExactLength<P4, 4>,
+              IsHexStringOfExactLength<P5, 12>
+          ] extends [true, true, true, true, true]
+            ? true
             : false
         : false
-
-    type AllElementsAreGuids<S extends readonly string[]> = S extends readonly [
-        infer F extends string,
-        ...infer R extends readonly string[]
-    ]
-        ? AssertGuid<F> extends never
-            ? false
-            : AllElementsAreGuids<R>
-        : true
 
     type AssertGuid<S extends string> = string extends S
         ? S
         : S extends `${infer P1}-${infer P2}-${infer P3}-${infer P4}-${infer P5}`
-        ? ValidateGuidParts<P1, P2, P3, P4, P5> extends true
+        ? ValidateGuidParts<[P1, P2, P3, P4, P5]> extends true
             ? S
             : never
         : never
 
-    type AssertGuids<S extends readonly string[]> = AllElementsAreGuids<S> extends true ? S : never
+    type AssertGuids<S extends readonly string[]> = {
+        [K in keyof S]: AssertGuid<S[K]>
+    }
 }